requirements.txt
numpy
scipy
joblib
pillow
pybind11